+++ sdk/storage/Azure.Storage.DataMovement/api/Azure.Storage.DataMovement.net6.0.cs
@@ -1,23 +1,5 @@
 namespace Azure.Storage.DataMovement
 {
-    public partial class DataTransferProperty
-    {
-        public DataTransferProperty() { }
-        public DataTransferProperty(bool preserve) { }
-        public virtual bool Preserve { get { throw null; } }
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
-        public override bool Equals(object? obj) { throw null; }
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
-        public override int GetHashCode() { throw null; }
-        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
-        public override string? ToString() { throw null; }
-    }
-    public partial class DataTransferProperty<T> : Azure.Storage.DataMovement.DataTransferProperty where T : notnull
-    {
-        public DataTransferProperty(bool preserve) { }
-        public DataTransferProperty(T value) { }
-        public virtual T? Value { get { throw null; } }
-    }
     public partial class LocalFilesStorageResourceProvider : Azure.Storage.DataMovement.StorageResourceProvider
     {
         public LocalFilesStorageResourceProvider() { }
@@ -27,24 +9,18 @@ namespace Azure.Storage.DataMovement
         public Azure.Storage.DataMovement.StorageResourceItem FromFile(string filePath) { throw null; }
         protected internal override System.Threading.Tasks.Task<Azure.Storage.DataMovement.StorageResource> FromSourceAsync(Azure.Storage.DataMovement.TransferProperties properties, System.Threading.CancellationToken cancellationToken) { throw null; }
     }
-    public partial class ProgressHandlerOptions
-    {
-        public ProgressHandlerOptions(System.IProgress<Azure.Storage.DataMovement.TransferProgress> progressHandler, bool trackBytesTransferred = false) { }
-        public System.IProgress<Azure.Storage.DataMovement.TransferProgress> ProgressHandler { get { throw null; } set { } }
-        public bool TrackBytesTransferred { get { throw null; } set { } }
-    }
     public abstract partial class StorageResource
     {
         protected StorageResource() { }
         protected internal abstract bool IsContainer { get; }
         public abstract string ProviderId { get; }
         public abstract System.Uri Uri { get; }
-        protected internal abstract Azure.Storage.DataMovement.StorageResourceCheckpointData GetDestinationCheckpointData();
-        protected internal abstract Azure.Storage.DataMovement.StorageResourceCheckpointData GetSourceCheckpointData();
+        protected internal abstract Azure.Storage.DataMovement.StorageResourceCheckpointDetails GetDestinationCheckpointDetails();
+        protected internal abstract Azure.Storage.DataMovement.StorageResourceCheckpointDetails GetSourceCheckpointDetails();
     }
-    public abstract partial class StorageResourceCheckpointData
+    public abstract partial class StorageResourceCheckpointDetails
     {
-        protected StorageResourceCheckpointData() { }
+        protected StorageResourceCheckpointDetails() { }
         public abstract int Length { get; }
         protected internal abstract void Serialize(System.IO.Stream stream);
     }
@@ -69,7 +45,7 @@ namespace Azure.Storage.DataMovement
         public Azure.HttpAuthorization SourceAuthentication { get { throw null; } set { } }
         public Azure.Storage.DataMovement.StorageResourceItemProperties SourceProperties { get { throw null; } set { } }
     }
-    public enum StorageResourceCreationPreference
+    public enum StorageResourceCreationMode
     {
         Default = 0,
         FailIfExists = 1,
@@ -99,11 +75,10 @@ namespace Azure.Storage.DataMovement
     }
     public partial class StorageResourceItemProperties
     {
-        protected StorageResourceItemProperties() { }
-        public StorageResourceItemProperties(long? resourceLength, Azure.ETag? eTag, System.DateTimeOffset? lastModifiedTime, System.Collections.Generic.Dictionary<string, object> properties) { }
+        public StorageResourceItemProperties() { }
         public Azure.ETag? ETag { get { throw null; } set { } }
         public System.DateTimeOffset? LastModifiedTime { get { throw null; } set { } }
-        public System.Collections.Generic.Dictionary<string, object> RawProperties { get { throw null; } set { } }
+        public System.Collections.Generic.IDictionary<string, object> RawProperties { get { throw null; } set { } }
         public long? ResourceLength { get { throw null; } set { } }
     }
     public abstract partial class StorageResourceProvider
@@ -132,8 +107,8 @@ namespace Azure.Storage.DataMovement
     public partial class TransferCheckpointStoreOptions
     {
         internal TransferCheckpointStoreOptions() { }
-        public static Azure.Storage.DataMovement.TransferCheckpointStoreOptions Disabled() { throw null; }
-        public static Azure.Storage.DataMovement.TransferCheckpointStoreOptions Local(string localCheckpointerPath) { throw null; }
+        public static Azure.Storage.DataMovement.TransferCheckpointStoreOptions CreateLocalStore(string localCheckpointPath) { throw null; }
+        public static Azure.Storage.DataMovement.TransferCheckpointStoreOptions DisableCheckpoint() { throw null; }
     }
     [System.FlagsAttribute]
     public enum TransferErrorMode
@@ -149,21 +124,21 @@ namespace Azure.Storage.DataMovement
     public partial class TransferItemCompletedEventArgs : Azure.Storage.DataMovement.TransferEventArgs
     {
         public TransferItemCompletedEventArgs(string transferId, Azure.Storage.DataMovement.StorageResourceItem sourceResource, Azure.Storage.DataMovement.StorageResourceItem destinationResource, bool isRunningSynchronously, System.Threading.CancellationToken cancellationToken) : base (default(string), default(bool), default(System.Threading.CancellationToken)) { }
-        public Azure.Storage.DataMovement.StorageResourceItem DestinationResource { get { throw null; } }
-        public Azure.Storage.DataMovement.StorageResourceItem SourceResource { get { throw null; } }
+        public Azure.Storage.DataMovement.StorageResourceItem Destination { get { throw null; } }
+        public Azure.Storage.DataMovement.StorageResourceItem Source { get { throw null; } }
     }
     public partial class TransferItemFailedEventArgs : Azure.Storage.DataMovement.TransferEventArgs
     {
         public TransferItemFailedEventArgs(string transferId, Azure.Storage.DataMovement.StorageResource sourceResource, Azure.Storage.DataMovement.StorageResource destinationResource, System.Exception exception, bool isRunningSynchronously, System.Threading.CancellationToken cancellationToken) : base (default(string), default(bool), default(System.Threading.CancellationToken)) { }
-        public Azure.Storage.DataMovement.StorageResource DestinationResource { get { throw null; } }
+        public Azure.Storage.DataMovement.StorageResource Destination { get { throw null; } }
         public System.Exception Exception { get { throw null; } }
-        public Azure.Storage.DataMovement.StorageResource SourceResource { get { throw null; } }
+        public Azure.Storage.DataMovement.StorageResource Source { get { throw null; } }
     }
     public partial class TransferItemSkippedEventArgs : Azure.Storage.DataMovement.TransferEventArgs
     {
         public TransferItemSkippedEventArgs(string transferId, Azure.Storage.DataMovement.StorageResourceItem sourceResource, Azure.Storage.DataMovement.StorageResourceItem destinationResource, bool isRunningSynchronously, System.Threading.CancellationToken cancellationToken) : base (default(string), default(bool), default(System.Threading.CancellationToken)) { }
-        public Azure.Storage.DataMovement.StorageResourceItem DestinationResource { get { throw null; } }
-        public Azure.Storage.DataMovement.StorageResourceItem SourceResource { get { throw null; } }
+        public Azure.Storage.DataMovement.StorageResourceItem Destination { get { throw null; } }
+        public Azure.Storage.DataMovement.StorageResourceItem Source { get { throw null; } }
     }
     public partial class TransferManager : System.IAsyncDisposable
     {
@@ -171,7 +146,7 @@ namespace Azure.Storage.DataMovement
         public TransferManager(Azure.Storage.DataMovement.TransferManagerOptions options = null) { }
         public virtual System.Collections.Generic.IAsyncEnumerable<Azure.Storage.DataMovement.TransferProperties> GetResumableTransfersAsync([System.Runtime.CompilerServices.EnumeratorCancellationAttribute] System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
         public virtual System.Collections.Generic.IAsyncEnumerable<Azure.Storage.DataMovement.TransferOperation> GetTransfersAsync(System.Collections.Generic.ICollection<Azure.Storage.DataMovement.TransferStatus> filterByStatus = null, [System.Runtime.CompilerServices.EnumeratorCancellationAttribute] System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-        public virtual System.Threading.Tasks.Task PauseTransferIfRunningAsync(string transferId, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
+        public virtual System.Threading.Tasks.Task PauseTransferAsync(string transferId, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
         public virtual System.Threading.Tasks.Task<System.Collections.Generic.List<Azure.Storage.DataMovement.TransferOperation>> ResumeAllTransfersAsync(Azure.Storage.DataMovement.TransferOptions transferOptions = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
         public virtual System.Threading.Tasks.Task<Azure.Storage.DataMovement.TransferOperation> ResumeTransferAsync(string transferId, Azure.Storage.DataMovement.TransferOptions transferOptions = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
         public virtual System.Threading.Tasks.Task<Azure.Storage.DataMovement.TransferOperation> StartTransferAsync(Azure.Storage.DataMovement.StorageResource sourceResource, Azure.Storage.DataMovement.StorageResource destinationResource, Azure.Storage.DataMovement.TransferOptions transferOptions = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
@@ -180,9 +155,9 @@ namespace Azure.Storage.DataMovement
     public partial class TransferManagerOptions
     {
         public TransferManagerOptions() { }
-        public Azure.Storage.DataMovement.TransferCheckpointStoreOptions CheckpointerOptions { get { throw null; } set { } }
+        public Azure.Storage.DataMovement.TransferCheckpointStoreOptions CheckpointStoreOptions { get { throw null; } set { } }
         public Azure.Core.DiagnosticsOptions Diagnostics { get { throw null; } }
-        public Azure.Storage.DataMovement.TransferErrorMode ErrorHandling { get { throw null; } set { } }
+        public Azure.Storage.DataMovement.TransferErrorMode ErrorMode { get { throw null; } set { } }
         public int? MaximumConcurrency { get { throw null; } set { } }
         public System.Collections.Generic.List<Azure.Storage.DataMovement.StorageResourceProvider> ResumeProviders { get { throw null; } set { } }
     }
@@ -199,10 +174,10 @@ namespace Azure.Storage.DataMovement
     public partial class TransferOptions : System.IEquatable<Azure.Storage.DataMovement.TransferOptions>
     {
         public TransferOptions() { }
-        public Azure.Storage.DataMovement.StorageResourceCreationPreference CreationPreference { get { throw null; } set { } }
+        public Azure.Storage.DataMovement.StorageResourceCreationMode CreationPreference { get { throw null; } set { } }
         public long? InitialTransferSize { get { throw null; } set { } }
         public long? MaximumTransferChunkSize { get { throw null; } set { } }
-        public Azure.Storage.DataMovement.ProgressHandlerOptions ProgressHandlerOptions { get { throw null; } set { } }
+        public Azure.Storage.DataMovement.TransferProgressHandlerOptions ProgressHandlerOptions { get { throw null; } set { } }
         public event Azure.Core.SyncAsyncEventHandler<Azure.Storage.DataMovement.TransferItemCompletedEventArgs> ItemTransferCompleted { add { } remove { } }
         public event Azure.Core.SyncAsyncEventHandler<Azure.Storage.DataMovement.TransferItemFailedEventArgs> ItemTransferFailed { add { } remove { } }
         public event Azure.Core.SyncAsyncEventHandler<Azure.Storage.DataMovement.TransferItemSkippedEventArgs> ItemTransferSkipped { add { } remove { } }
@@ -233,14 +208,20 @@ namespace Azure.Storage.DataMovement
         public long QueuedCount { get { throw null; } }
         public long SkippedCount { get { throw null; } }
     }
+    public partial class TransferProgressHandlerOptions
+    {
+        public TransferProgressHandlerOptions() { }
+        public System.IProgress<Azure.Storage.DataMovement.TransferProgress> ProgressHandler { get { throw null; } set { } }
+        public bool TrackBytesTransferred { get { throw null; } set { } }
+    }
     public partial class TransferProperties
     {
         protected internal TransferProperties() { }
-        public virtual byte[] DestinationCheckpointData { get { throw null; } }
+        public virtual byte[] DestinationCheckpointDetails { get { throw null; } }
         public virtual string DestinationProviderId { get { throw null; } }
         public virtual System.Uri DestinationUri { get { throw null; } }
         public virtual bool IsContainer { get { throw null; } }
-        public virtual byte[] SourceCheckpointData { get { throw null; } }
+        public virtual byte[] SourceCheckpointDetails { get { throw null; } }
         public virtual string SourceProviderId { get { throw null; } }
         public virtual System.Uri SourceUri { get { throw null; } }
         public virtual string TransferId { get { throw null; } }