+++ sdk/storage/Azure.Storage.DataMovement/src/JobPlan/JobPlanHeader.cs
@@ -69,16 +69,16 @@ namespace Azure.Storage.DataMovement.JobPlan
         /// Additional checkpoint data specific to the source resource.
         /// Only populated when using <see cref="Deserialize(Stream)"/>.
         /// </summary>
-        public byte[] SourceCheckpointData;
+        public byte[] SourceCheckpointDetails;
 
         /// <summary>
         /// Additional checkpoint data specific to the destination resource.
         /// Only populated when using <see cref="Deserialize(Stream)"/>.
         /// </summary>
-        public byte[] DestinationCheckpointData;
+        public byte[] DestinationCheckpointDetails;
 
-        private StorageResourceCheckpointData _sourceCheckpointData;
-        private StorageResourceCheckpointData _destinationCheckpointData;
+        private StorageResourceCheckpointDetails _sourceCheckpointDetails;
+        private StorageResourceCheckpointDetails _destinationCheckpointDetails;
 
         public JobPlanHeader(
             int version,
@@ -92,8 +92,8 @@ namespace Azure.Storage.DataMovement.JobPlan
             TransferStatus jobStatus,
             string parentSourcePath,
             string parentDestinationPath,
-            StorageResourceCheckpointData sourceCheckpointData,
-            StorageResourceCheckpointData destinationCheckpointData)
+            StorageResourceCheckpointDetails sourceCheckpointDetails,
+            StorageResourceCheckpointDetails destinationCheckpointDetails)
         {
             Argument.AssertNotNullOrEmpty(transferId, nameof(transferId));
             Argument.AssertNotNullOrEmpty(sourceProviderId, nameof(sourceProviderId));
@@ -102,8 +102,8 @@ namespace Azure.Storage.DataMovement.JobPlan
             Argument.AssertNotNull(createTime, nameof(createTime));
             Argument.AssertNotNullOrEmpty(parentSourcePath, nameof(parentSourcePath));
             Argument.AssertNotNullOrEmpty(parentDestinationPath, nameof(parentDestinationPath));
-            Argument.AssertNotNull(sourceCheckpointData, nameof(sourceCheckpointData));
-            Argument.AssertNotNull(destinationCheckpointData, nameof(destinationCheckpointData));
+            Argument.AssertNotNull(sourceCheckpointDetails, nameof(sourceCheckpointDetails));
+            Argument.AssertNotNull(destinationCheckpointDetails, nameof(destinationCheckpointDetails));
 
             if (sourceProviderId.Length > DataMovementConstants.JobPlanFile.ProviderIdMaxLength)
             {
@@ -126,8 +126,8 @@ namespace Azure.Storage.DataMovement.JobPlan
             ParentSourcePath = parentSourcePath;
             ParentDestinationPath = parentDestinationPath;
 
-            _sourceCheckpointData = sourceCheckpointData;
-            _destinationCheckpointData = destinationCheckpointData;
+            _sourceCheckpointDetails = sourceCheckpointDetails;
+            _destinationCheckpointDetails = destinationCheckpointDetails;
         }
 
         private JobPlanHeader(
@@ -142,8 +142,8 @@ namespace Azure.Storage.DataMovement.JobPlan
             TransferStatus jobStatus,
             string parentSourcePath,
             string parentDestinationPath,
-            byte[] sourceCheckpointData,
-            byte[] destinationCheckpointData)
+            byte[] sourceCheckpointDetails,
+            byte[] destinationCheckpointDetails)
         {
             Version = version;
             TransferId = transferId;
@@ -156,8 +156,8 @@ namespace Azure.Storage.DataMovement.JobPlan
             JobStatus = jobStatus;
             ParentSourcePath = parentSourcePath;
             ParentDestinationPath = parentDestinationPath;
-            SourceCheckpointData = sourceCheckpointData;
-            DestinationCheckpointData = destinationCheckpointData;
+            SourceCheckpointDetails = sourceCheckpointDetails;
+            DestinationCheckpointDetails = destinationCheckpointDetails;
         }
 
         public void Serialize(Stream stream)
@@ -203,11 +203,11 @@ namespace Azure.Storage.DataMovement.JobPlan
             byte[] parentDestinationPathBytes = Encoding.UTF8.GetBytes(ParentDestinationPath);
             writer.WriteVariableLengthFieldInfo(parentDestinationPathBytes.Length, ref currentVariableLengthIndex);
 
-            // SourceCheckpointData offset/length
-            writer.WriteVariableLengthFieldInfo(_sourceCheckpointData.Length, ref currentVariableLengthIndex);
+            // SourceCheckpointDetails offset/length
+            writer.WriteVariableLengthFieldInfo(_sourceCheckpointDetails.Length, ref currentVariableLengthIndex);
 
-            // DestinationCheckpointData offset/length
-            writer.WriteVariableLengthFieldInfo(_destinationCheckpointData.Length, ref currentVariableLengthIndex);
+            // DestinationCheckpointDetails offset/length
+            writer.WriteVariableLengthFieldInfo(_destinationCheckpointDetails.Length, ref currentVariableLengthIndex);
 
             // ParentSourcePath
             writer.Write(parentSourcePathBytes);
@@ -215,8 +215,8 @@ namespace Azure.Storage.DataMovement.JobPlan
             // ParentDestinationPath
             writer.Write(parentDestinationPathBytes);
 
-            _sourceCheckpointData.Serialize(stream);
-            _destinationCheckpointData.Serialize(stream);
+            _sourceCheckpointDetails.Serialize(stream);
+            _destinationCheckpointDetails.Serialize(stream);
         }
 
         public static JobPlanHeader Deserialize(Stream stream)
@@ -270,13 +270,13 @@ namespace Azure.Storage.DataMovement.JobPlan
             int parentDestinationPathOffset = reader.ReadInt32();
             int parentDestinationPathLength = reader.ReadInt32();
 
-            // SourceCheckpointData offset/length
-            int sourceCheckpointDataOffset = reader.ReadInt32();
-            int sourceCheckpointDataLength = reader.ReadInt32();
+            // SourceCheckpointDetails offset/length
+            int sourceCheckpointDetailsOffset = reader.ReadInt32();
+            int sourceCheckpointDetailsLength = reader.ReadInt32();
 
-            // DestinationCheckpointData offset/length
-            int destinationCheckpointDataOffset = reader.ReadInt32();
-            int destinationCheckpointDataLength = reader.ReadInt32();
+            // DestinationCheckpointDetails offset/length
+            int destinationCheckpointDetailsOffset = reader.ReadInt32();
+            int destinationCheckpointDetailsLength = reader.ReadInt32();
 
             // ParentSourcePath
             string parentSourcePath = null;
@@ -296,20 +296,20 @@ namespace Azure.Storage.DataMovement.JobPlan
                 parentDestinationPath = parentDestinationPathBytes.ToString(parentDestinationPathLength);
             }
 
-            // SourceCheckpointData
-            byte[] sourceCheckpointData = Array.Empty<byte>();
-            if (sourceCheckpointDataOffset > 0)
+            // SourceCheckpointDetails
+            byte[] sourceCheckpointDetails = Array.Empty<byte>();
+            if (sourceCheckpointDetailsOffset > 0)
             {
-                reader.BaseStream.Position = sourceCheckpointDataOffset;
-                sourceCheckpointData = reader.ReadBytes(sourceCheckpointDataLength);
+                reader.BaseStream.Position = sourceCheckpointDetailsOffset;
+                sourceCheckpointDetails = reader.ReadBytes(sourceCheckpointDetailsLength);
             }
 
-            // DestinationCheckpointData
-            byte[] destinationCheckpointData = Array.Empty<byte>();
-            if (destinationCheckpointDataOffset > 0)
+            // DestinationCheckpointDetails
+            byte[] destinationCheckpointDetails = Array.Empty<byte>();
+            if (destinationCheckpointDetailsOffset > 0)
             {
-                reader.BaseStream.Position = destinationCheckpointDataOffset;
-                destinationCheckpointData = reader.ReadBytes(destinationCheckpointDataLength);
+                reader.BaseStream.Position = destinationCheckpointDetailsOffset;
+                destinationCheckpointDetails = reader.ReadBytes(destinationCheckpointDetailsLength);
             }
 
             return new JobPlanHeader(
@@ -324,8 +324,8 @@ namespace Azure.Storage.DataMovement.JobPlan
                 jobPlanStatus.ToTransferStatus(),
                 parentSourcePath,
                 parentDestinationPath,
-                sourceCheckpointData,
-                destinationCheckpointData);
+                sourceCheckpointDetails,
+                destinationCheckpointDetails);
         }
     }
 }